feat(cloud): Moved usageProcessorTrigger to separate functions app
