+++ cloud/src/Signal.Api.sln
@@ -24,7 +24,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Signal.Api.Internal", "Sign
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Signal.Api.Common", "Signal.Api.Common\Signal.Api.Common.csproj", "{6187DDDC-013E-4226-942B-A6EBAF48EE6A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Signalco.Cloud.Channel.GitHubApp", "Signalco.Cloud.Channel.GitHubApp\Signalco.Cloud.Channel.GitHubApp.csproj", "{EE218B89-6F1E-4CF1-AF27-E530B6419951}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Signalco.Cloud.Channel.GitHubApp", "Signalco.Cloud.Channel.GitHubApp\Signalco.Cloud.Channel.GitHubApp.csproj", "{EE218B89-6F1E-4CF1-AF27-E530B6419951}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Signalco.Channel.Slack", "Signalco.Channel.Slack\Signalco.Channel.Slack.csproj", "{A8B25F97-1EDB-4D36-8003-2AE6D604C684}"
 EndProject
@@ -46,7 +46,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_Shared", "_Shared", "{88AB
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_API", "_API", "{22B3C3FC-54BE-4990-8BB6-55141C315A04}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Signalco.Api.Common.Health", "Signalco.Api.Common.Health\Signalco.Api.Common.Health.csproj", "{308028B9-103E-4C6E-B4BE-DF9707F130FB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Signalco.Api.Common.Health", "Signalco.Api.Common.Health\Signalco.Api.Common.Health.csproj", "{308028B9-103E-4C6E-B4BE-DF9707F130FB}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Internal", "Internal", "{F8CEDDCF-4DC6-46FE-9F9B-D9AB60C147B6}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Signalco.Func.Internal.UsageProcessor", "Signalco.Func.Internal.UsageProcessor\Signalco.Func.Internal.UsageProcessor.csproj", "{B2BC2CDB-8022-4E81-AD94-7B136796992F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -118,6 +122,10 @@ Global
 		{308028B9-103E-4C6E-B4BE-DF9707F130FB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{308028B9-103E-4C6E-B4BE-DF9707F130FB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{308028B9-103E-4C6E-B4BE-DF9707F130FB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B2BC2CDB-8022-4E81-AD94-7B136796992F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B2BC2CDB-8022-4E81-AD94-7B136796992F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B2BC2CDB-8022-4E81-AD94-7B136796992F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B2BC2CDB-8022-4E81-AD94-7B136796992F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -131,6 +139,7 @@ Global
 		{94C41557-A004-491A-AF09-53E47DB07E92} = {22B3C3FC-54BE-4990-8BB6-55141C315A04}
 		{94B4DC92-36CC-4F71-B12A-4AE2CCB1228C} = {22B3C3FC-54BE-4990-8BB6-55141C315A04}
 		{6187DDDC-013E-4226-942B-A6EBAF48EE6A} = {88AB2564-9033-4289-9101-BF26E56F7595}
+		{EE218B89-6F1E-4CF1-AF27-E530B6419951} = {DCF287D4-46DA-4E27-80CD-34EEDC3766A4}
 		{A8B25F97-1EDB-4D36-8003-2AE6D604C684} = {DCF287D4-46DA-4E27-80CD-34EEDC3766A4}
 		{080869D1-9C97-4678-93A9-0197EDB87592} = {DCF287D4-46DA-4E27-80CD-34EEDC3766A4}
 		{9AC9A8B0-F7AF-4CF9-9C76-2EC883EE2737} = {DCF287D4-46DA-4E27-80CD-34EEDC3766A4}
@@ -138,6 +147,7 @@ Global
 		{98D9F15A-DA05-4F5B-AD39-7744C24AC149} = {48221AEC-79B1-4CB9-BCFD-5424BAA29D67}
 		{E87409AB-7BCC-4501-9882-7F2004283896} = {DCF287D4-46DA-4E27-80CD-34EEDC3766A4}
 		{308028B9-103E-4C6E-B4BE-DF9707F130FB} = {88AB2564-9033-4289-9101-BF26E56F7595}
+		{B2BC2CDB-8022-4E81-AD94-7B136796992F} = {F8CEDDCF-4DC6-46FE-9F9B-D9AB60C147B6}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {46B06AB8-117A-49FE-92EE-CBD00E067023}
