+++ cloud/src/Signalco.Func.Internal.UsageProcessor/host.json
@@ -11,7 +11,7 @@
   },
   "extensions": {
     "queues": {
-      "maxPollingInterval": "00:00:10"
+      "maxPollingInterval": "00:10:00"
     }
   }
 }
